Test return value of writer methods
These is a special case: the method itself is treated as a void method,
and the assigned value is what is actually used. These methods don't
really return anything, and should not be validated. (One might even
argue that they should not be documented, which fixes this test as
well).

Something similar happens in the initialize, but that one completely
breaks the tests.